refactor: move htn planning under missions
